requirements.txt
click==6.7
colormath==3.0.0
colorthief==0.2.1
decorator==4.2.1
dlib==19.8.1
face-recognition==1.2.1
face-recognition-models==0.3.0
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
networkx==2.1
numpy==1.14.2
opencv-python==3.4.0.12
Pillow==5.0.0
pip==9.0.1
pymongo==3.6.1
scipy==1.0.0
setuptools==28.8.0
Werkzeug==0.14.1
